requirements.txt
# ================================
# Multi-Agent Research System
# ================================

# Core agent framework
crewai==1.7.2

# LLM provider
litellm==1.80.11

# Search tools
tavily-python==0.7.21
duckduckgo-search==8.1.1

# Wikipedia search
wikipedia-api==0.10.0

# HTTP requests (model testing)
requests==2.32.5

# Environment variables
python-dotenv==1.1.1

# Typing support
typing_extensions>=4.9.0



